+++ src/lib/meeting-card/mapper.ts
@@ -64,62 +64,6 @@ const formatTime = (time: string) => {
 	});
 };
 
-export type MeetingWithDates = Pick<
-	SelectMeeting,
-	"id" | "scheduled" | "dates"
-> & {
-	needsAvailability: boolean;
-	allAvailabilityFilled: boolean;
-	hostId: string;
-};
-
-export function getStartOfTodayMs(): number {
-	const d = new Date();
-	d.setHours(0, 0, 0, 0);
-	return d.getTime();
-}
-
-export function isMeetingPast(
-	m: Pick<SelectMeeting, "id" | "scheduled" | "dates">,
-	scheduledDates: Record<string, number> | undefined,
-	todayMs: number,
-): boolean {
-	if (m.scheduled) {
-		const scheduledDate = scheduledDates?.[m.id];
-		if (scheduledDate === undefined) return false;
-		return scheduledDate < todayMs;
-	}
-	const dates = (m.dates as string[] | null) ?? [];
-	if (dates.length === 0) return false;
-	const lastDate = Math.max(...dates.map((d) => new Date(d).getTime()));
-	return lastDate < todayMs;
-}
-
-export function getMeetingUpcomingPriority(
-	m: MeetingWithDates,
-	memberId: string,
-	upcomingSet: Set<string>,
-): number {
-	if (m.needsAvailability) return 0;
-	if (m.allAvailabilityFilled && m.hostId === memberId) return 1;
-	if (upcomingSet.has(m.id)) return 2;
-	if (m.scheduled) return 3;
-	return 4;
-}
-
-export function filterMeetingsByQuery<
-	T extends Pick<SelectMeeting, "title" | "location" | "description">,
->(meetings: T[], query: string): T[] {
-	const normalized = query.trim().toLowerCase();
-	if (!normalized) return meetings;
-	return meetings.filter(
-		(m) =>
-			m.title.toLowerCase().includes(normalized) ||
-			(m.location ?? "").toLowerCase().includes(normalized) ||
-			(m.description ?? "").toLowerCase().includes(normalized),
-	);
-}
-
 export function toMeetingCardData(
 	meeting: MeetingForCard,
 	memberId: string,
